+++ bitorch_engine/utils/cuda_extension.py
@@ -1,5 +1,6 @@
 import os
 import subprocess
+import warnings
 from pathlib import Path
 from typing import Dict, Any
 
@@ -8,7 +9,19 @@ from torch.utils.cpp_extension import CUDAExtension
 from bitorch_engine.extensions import EXTENSION_PREFIX
 
 
+SUPPORTED_CUDA_ARCHS = ["sm_80", "sm_75"]
+
+
+def get_cuda_arch():
+    cuda_arch = os.environ.get("BIE_CUDA_ARCH", "sm_80")
+    if cuda_arch not in SUPPORTED_CUDA_ARCHS:
+        warnings.warn(f"Warning: BIE_CUDA_ARCH={cuda_arch} may not be supported yet.")
+    return cuda_arch
+
+
 def get_cuda_extension(root_path: Path, relative_name: str, relative_sources) -> Any:
+    if os.environ.get("BIE_BUILD_SEPARATE_CUDA_ARCH", "false") == "true":
+        relative_name = f"{relative_name}-{get_cuda_arch()}"
     return CUDAExtension(
         name=EXTENSION_PREFIX + relative_name,
         sources=[str(root_path / rel_path) for rel_path in relative_sources],
@@ -65,20 +78,21 @@ def get_kwargs() -> Dict[str, Any]:
     kwargs = {
         "extra_compile_args": {
             "cxx": [
-                "-Wno-deprecated-declarations", # Suppress warnings for deprecated declarations
-                "-L/usr/lib/gcc/x86_64-pc-linux-gnu/10.3.0/libgomp.so", # Specify libgomp library path for OpenMP
+                "-Wno-deprecated-declarations",  # Suppress warnings for deprecated declarations
+                "-L/usr/lib/gcc/x86_64-pc-linux-gnu/10.3.0/libgomp.so",  # Specify libgomp library path for OpenMP
                 "-fopenmp", # Enable OpenMP support
             ],
             "nvcc": [
                 "-Xcompiler",
-                "-fopenmp", # Pass -fopenmp to the host compiler via nvcc
-                f"-arch={os.environ.get('BIE_CUDA_ARCH', 'sm_80')}", # Set CUDA architecture, default to 'sm_80'
-                "-DARCH_SM_75" if os.environ.get('BIE_CUDA_ARCH') == 'sm_75' else "-DARCH_SM_80", # Define architecture macro based on BIE_CUDA_ARCH
+                "-fopenmp",  # Pass -fopenmp to the host compiler via nvcc
+                f"-arch={get_cuda_arch()}",  # Set CUDA architecture, default to 'sm_80'
+                "-DARCH_SM_75" if get_cuda_arch() == 'sm_75' else "-DARCH_SM_80",  # set architecture macro
             ],
         },
     }
     # If the GCC version is greater than 11, adjust the nvcc host compiler settings
     if major > 11:
         print("Using GCC 11 host compiler for nvcc.")
-        kwargs["extra_compile_args"]["nvcc"].append("-ccbin=/usr/bin/gcc-11") # Specify GCC 11 as the host compiler for nvcc
+        # Specify GCC 11 as the host compiler for nvcc:
+        kwargs["extra_compile_args"]["nvcc"].append("-ccbin=/usr/bin/gcc-11")
     return kwargs